api/pyproject.toml
[tool.poetry]
name = "fake-api"
version = "0.1.0"
description = "API REST construída com FastAPI que fornece endpoints para produtos, carrinhos, clientes e logística"
authors = ["Go.nine <contato@gonine.io>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.104.1"
uvicorn = "0.24.0"
sqlalchemy = "2.0.23"
pyjwt = "2.8.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-multipart = "0.0.6"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
faker = "^33.3.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 